feat(layout): 添加智能体相关页面和布局组件
- 新增 `AgentLayout` 作为主布局组件，包含侧边菜单和历史会话功能。
- 创建 `AgentChatpage` 和 `AgentHistoryConversation` 页面，分别用于智能体聊天和历史会话展示。
- 引入样式文件 `index.less`，定义布局和组件样式。
- 更新路由配置，支持新的智能体相关页面。
- 优化 `AgentSidebar` 和 `AgentConversation` 组件，增加 `onViewMore` 回调以支持历史会话查看功能。
