+++ src/components/AgentSidebar/index.tsx
@@ -30,6 +30,7 @@ const AgentSidebar = forwardRef<AgentSidebarRef, AgentSidebarProps>(
       agentDetail,
       onToggleCollectSuccess,
       onVisibleChange,
+      onViewMore,
     },
     ref,
   ) => {
@@ -110,7 +111,10 @@ const AgentSidebar = forwardRef<AgentSidebarRef, AgentSidebarProps>(
                       onToggleCollectSuccess={onToggleCollectSuccess}
                     />
                     {/* 智能体相关会话 */}
-                    <AgentConversation agentId={agentId} />
+                    <AgentConversation
+                      agentId={agentId}
+                      onViewMore={onViewMore}
+                    />
                     {/* 定时任务 */}
                     {agentDetail?.openScheduledTask === OpenCloseEnum.Open && (
                       <TimedTask agentId={agentId} />
