+++ src/pages/Agent/AgentChatpage/index.tsx
@@ -1,0 +1,12 @@
+import Chat from '@/pages/Chat';
+import React from 'react';
+
+/**
+ * 智能体聊天页面
+ * @returns
+ */
+const AgentChatpage: React.FC = () => {
+  return <Chat />;
+};
+
+export default AgentChatpage;
