+++ src/pages/Agent/AgentHistoryConversation/index.tsx
@@ -1,0 +1,19 @@
+import HistoryConversationPage from '@/pages/HistoryConversation/components/HistoryConversationPage';
+import React from 'react';
+import { history, useParams } from 'umi';
+
+/**
+ * 历史会话页面（用于无菜单layouts的页面，单个智能体推荐使用，类似kimi的history页面）
+ */
+const AgentHistoryConversation: React.FC = () => {
+  const params = useParams();
+  const agentId = Number(params.agentId);
+
+  const handleLink = (id: number, agentId: number) => {
+    history.push(`/agent/chatpage/${id}/${agentId}`);
+  };
+
+  return <HistoryConversationPage agentId={agentId} onClickLink={handleLink} />;
+};
+
+export default AgentHistoryConversation;
