+++ src/pages/AgentDetails/AgentDetailsPage/index.less
